pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "risnaberti-nanumber"
version = "0.1.0"
description = "Universal auto-number generator by Risna Berti (Nanumber)"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
  { name = "Risna Berti", email = "risnaberti@gmail.com" }
]

keywords = ["autonumber", "generator", "sqlalchemy", "reset", "python", "nanumber"]
dependencies = [
  "SQLAlchemy>=1.4",
  "PyYAML>=6.0",
  "pydantic>=1.10"
]

[tool.setuptools.packages.find]
where = ["risnaberti"]
